% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{depricated}
\alias{depricated}
\alias{new_menu}
\title{Deprecated functions}
\usage{
new_menu(
  key = character(),
  value = list(),
  attrs = NULL,
  ...,
  class = character()
)
}
\description{
Deprecated functions
}
