% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item.R
\name{is_item}
\alias{is_item}
\title{Test if an object is an item}
\usage{
is_item(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{TRUE} if the object inherits from the \code{navigatr_item} class.
}
\description{
Test if an object is an item
}
