% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menu.R
\name{is_menu}
\alias{is_menu}
\title{Test if the object is a menu}
\usage{
is_menu(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{TRUE} if the object inherits from the \code{navigatr_menu} class.
}
\description{
Test if the object is a menu
}
