% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nav_description.R
\name{nav_description}
\alias{nav_description}
\alias{nav_description<-}
\alias{set_nav_description}
\title{Get or set navigation description}
\usage{
nav_description(x)

nav_description(x) <- value

set_nav_description(x, value)
}
\arguments{
\item{x}{An object to get or set the description for.}

\item{value}{A character string to set as the description.}
}
\value{
\code{nav_description()} returns the description of the object, or \code{NULL}
if no description is set. \verb{nav_description<-} and \code{set_nav_description()}
return the modified object.
}
\description{
These functions get or set descriptions of items in menus and input forms.
Description is used to print a description of each item in the menu or form.
If no description is set, the result of \code{\link[pillar:type_sum]{pillar::obj_sum()}} is used for the
description.
}
