% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navigatr-package.R
\docType{package}
\name{navigatr-package}
\alias{navigatr}
\alias{navigatr-package}
\title{navigatr: Navigation Menu for Pipe-Friendly Data Processing}
\description{
Provides a navigation menu to enable pipe-friendly data processing for hierarchical data structures. By activating the menu items, you can perform operations on each item while maintaining the overall structure in attributes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/UchidaMizuki/navigatr}
  \item \url{https://uchidamizuki.github.io/navigatr/}
  \item Report bugs at \url{https://github.com/UchidaMizuki/navigatr/issues}
}

}
\author{
\strong{Maintainer}: Mizuki Uchida \email{uchidamizuki@vivaldi.net}

}
\keyword{internal}
