% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menu.R
\name{new_menu}
\alias{new_menu}
\title{Build a new navigation menu}
\usage{
new_menu(
  key = character(),
  value = list(),
  attrs = data_frame(.size = 1),
  ...,
  class = character()
)
}
\arguments{
\item{key}{A unique character vector.
If the key is not a character vector, it is converted to a character vector by \code{\link[=as.character]{as.character()}}.}

\item{value}{A list of values corresponding to the keys.
If the value is not a list, it is converted to a list by \code{\link[=as.list]{as.list()}}.}

\item{attrs}{A data frame for additional attributes of items (an empty data frame by default).
When an item becomes active, the attrs will be added to its attributes.}

\item{...}{Additional arguments passed to \code{\link[vctrs:new_data_frame]{vctrs::new_data_frame()}}.}

\item{class}{A character vector of subclasses passed to \code{\link[vctrs:new_data_frame]{vctrs::new_data_frame()}}.}
}
\value{
A \code{menu} object, a subclass of class \code{data.frame}.
}
\description{
\code{new_menu()} build a new navigation menu.
The menu has unique keys and their corresponding values.
Each menu item can be accessed by \code{\link[=activate]{activate()}}.
}
\examples{
library(dplyr)

mn1 <- new_menu(key = c("band_members", "band_instruments"),
                value = list(band_members, band_instruments))
mn1

# You can also build a nested menu
mn2 <- new_menu(key = c("key1", "key2"),
                value = list(mn1, mn1))
mn2

}
\seealso{
\code{\link[=activate]{activate()}}
}
