% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item.R
\name{rekey}
\alias{rekey}
\title{Rename a key}
\usage{
rekey(data, new_key)
}
\arguments{
\item{data}{A \code{navigatr_item} object.}

\item{new_key}{New key name.}
}
\value{
A \code{navigatr_item} object.
}
\description{
Rename a key
}
