% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemise.R
\name{itemise}
\alias{itemise}
\alias{itemize}
\alias{itemise.navigatr_nav}
\alias{itemise.navigatr_item}
\title{Update items}
\usage{
itemise(.data, ...)

itemize(.data, ...)

\method{itemise}{navigatr_nav}(.data, ...)

\method{itemise}{navigatr_item}(.data, ...)
}
\arguments{
\item{.data}{A \code{navigatr_nav} or \code{navigatr_item} object.}

\item{...}{For \code{navigatr_nav} objects, use \code{key_name = item_value}. For
\code{navigatr_item} objects, new item value.}
}
\value{
A \code{navigatr_nav} or \code{navigatr_item} object.
}
\description{
Update items
}
