% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemise.R
\name{itemise}
\alias{itemise}
\alias{itemize}
\title{Update items}
\usage{
itemise(.data, ...)

itemize(.data, ...)
}
\arguments{
\item{.data}{A \code{navigatr_nav_input} object.}

\item{...}{Key-value pairs.}
}
\value{
A \code{navigatr_nav_input} object.
}
\description{
Update items
}
