% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{new_menu}
\alias{new_menu}
\title{Deprecated functions}
\usage{
new_menu(
  key = character(),
  value = list(),
  attrs = NULL,
  ...,
  class = character()
)
}
\arguments{
\item{key}{A unique character vector.}

\item{value}{A list of values corresponding to the keys.}

\item{attrs}{A data frame for additional attributes of items (an empty data frame by default).
When an item becomes active, the attrs will be added to its attributes.}

\item{...}{Additional arguments passed to \code{\link[vctrs:new_data_frame]{vctrs::new_data_frame()}}.}

\item{class}{A character vector of subclasses passed to \code{\link[vctrs:new_data_frame]{vctrs::new_data_frame()}}.}
}
\value{
A \code{navigatr_nav_menu} object, a subclass of class \code{data.frame}.
}
\description{
Deprecated functions
}
