% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rekey.R
\name{rekey}
\alias{rekey}
\alias{rekey.navigatr_menu}
\alias{rekey.navigatr_item}
\title{Rename key names}
\usage{
rekey(.data, ...)

\method{rekey}{navigatr_menu}(.data, ...)

\method{rekey}{navigatr_item}(.data, ...)
}
\arguments{
\item{.data}{A \code{navigatr_menu} or \code{navigatr_item} object.}

\item{...}{For \code{navigatr_menu} objects, use \code{new_name = old_name}. For
\code{navigatr_item} objects, a scalar character of the new key name.}
}
\value{
A \code{navigatr_menu} or \code{navigatr_item} object.
}
\description{
Rename key names
}
